Changed: #825 Remove all warning when compiling Ryzom on Linux

+++ code/ryzom/tools/leveldesign/georges_convert/form_body_elt_struct.cpp
@@ -162,7 +162,7 @@ bool CFormBodyEltStruct::Empty() const
 
 uint32 CFormBodyEltStruct::GetNbElt () const
 {
-	return vpbodyelt.size();
+	return (uint32)vpbodyelt.size();
 }
 
 CFormBodyElt* CFormBodyEltStruct::GetElt( const unsigned int _index ) const
